documentation_slack_bot/pyproject.toml
[project]
name = "documentation_slack_bot"
version = "0.1.0"
description = "documentation-slack-bot using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[project.scripts]
documentation_slack_bot = "documentation_slack_bot.main:run"
run_crew = "documentation_slack_bot.main:run"
train = "documentation_slack_bot.main:train"
replay = "documentation_slack_bot.main:replay"
test = "documentation_slack_bot.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
